Split NetHammer to separate package Updated validation to be static
